+++ src/client/components/EditorExtended/model/content.js
@@ -1,6 +1,40 @@
-import { EditorState, convertFromRaw, ContentState } from 'draft-js';
+import { CompositeDecorator, ContentState, convertFromRaw, EditorState } from 'draft-js';
+import ObjectLink, { findObjEntities } from '../components/entities/objectlink';
+import Link, { findLinkEntities } from '../components/entities/link';
+import ImageSideButton from '../components/sides/ImageSideButton';
+import VideoSideButton from '../components/sides/VideoSideButton';
+import ObjectSideButton from '../components/sides/ObjectSideButton';
+import SeparatorButton from '../components/sides/SeparatorSideButton';
 
-import { defaultDecorators } from '../../../helpers/editorHelper';
+export const defaultDecorators = new CompositeDecorator([
+  {
+    strategy: findObjEntities,
+    component: ObjectLink,
+  },
+  {
+    strategy: findLinkEntities,
+    component: Link,
+  },
+]);
+
+export const SIDE_BUTTONS = [
+  {
+    title: 'Image',
+    component: ImageSideButton,
+  },
+  {
+    title: 'Video',
+    component: VideoSideButton,
+  },
+  {
+    title: 'Object',
+    component: ObjectSideButton,
+  },
+  {
+    title: 'Separator',
+    component: SeparatorButton,
+  },
+];
 
 const createEditorState = (content = null, decorators = defaultDecorators) => {
   let initialEditorState = {};
