+++ src/client/components/EditorExtended/util/editorStateToMarkdown.js
@@ -1,3 +1,5 @@
+import { get } from 'lodash';
+
 import { ATOMIC_TYPES, Block, Entity } from './constants';
 
 const defaultMarkdownDict = {
@@ -142,91 +144,87 @@ function editorStateToMarkdown(raw, extraMarkdownDict) {
 
   let isListType = false;
 
-  raw &&
-    raw.blocks.forEach((block, blockIndex) => {
-      if (blockIndex !== 0) {
+  get(raw, 'blocks', []).forEach((block, blockIndex) => {
+    if (blockIndex !== 0) {
+      returnString += '\n';
+      totalOffset = 0;
+      if (
+        block.text &&
+        block.type !== 'ordered-list-item' &&
+        block.type !== 'unordered-list-item' &&
+        isListType
+      ) {
         returnString += '\n';
-        totalOffset = 0;
-        if (
-          block.text &&
-          block.type !== 'ordered-list-item' &&
-          block.type !== 'unordered-list-item' &&
-          isListType
-        ) {
-          returnString += '\n';
-        }
       }
-      isListType = block.type === 'ordered-list-item' || block.type === 'unordered-list-item';
+    }
+    isListType = block.type === 'ordered-list-item' || block.type === 'unordered-list-item';
 
-      // add block style
-      returnString += getBlockStyle(block.type, appliedBlockStyles);
-      appliedBlockStyles.push(block.type);
+    // add block style
+    returnString += getBlockStyle(block.type, appliedBlockStyles);
+    appliedBlockStyles.push(block.type);
 
-      const appliedStyles = [];
+    const appliedStyles = [];
 
-      returnString += block.text.split('').reduce((text, currentChar, index) => {
-        let newText = text;
-        const sortedInlineStyleRanges = getInlineStyleRangesByLength(block.inlineStyleRanges);
+    returnString += block.text.split('').reduce((text, currentChar, index) => {
+      let newText = text;
+      const sortedInlineStyleRanges = getInlineStyleRangesByLength(block.inlineStyleRanges);
 
-        // find all styled at this character
-        const stylesStartAtChar = sortedInlineStyleRanges
-          .filter(range => range.offset === index)
-          .filter(range => markdownDict[range.style]); // disregard styles not defined in the md dict
+      // find all styled at this character
+      const stylesStartAtChar = sortedInlineStyleRanges
+        .filter(range => range.offset === index)
+        .filter(range => markdownDict[range.style]); // disregard styles not defined in the md dict
 
-        // add the symbol to the md string and push the style in the applied styles stack
-        stylesStartAtChar.forEach(currentStyle => {
-          const symbolLength = markdownDict[currentStyle.style].length;
+      // add the symbol to the md string and push the style in the applied styles stack
+      stylesStartAtChar.forEach(currentStyle => {
+        const symbolLength = markdownDict[currentStyle.style].length;
 
-          newText += markdownDict[currentStyle.style];
-          totalOffset += symbolLength;
-          appliedStyles.push({
-            symbol: markdownDict[currentStyle.style],
-            range: {
-              start: currentStyle.offset + totalOffset,
-              end: currentStyle.offset + currentStyle.length + totalOffset,
-            },
-            end: currentStyle.offset + (currentStyle.length - 1),
-          });
+        newText += markdownDict[currentStyle.style];
+        totalOffset += symbolLength;
+        appliedStyles.push({
+          symbol: markdownDict[currentStyle.style],
+          range: {
+            start: currentStyle.offset + totalOffset,
+            end: currentStyle.offset + currentStyle.length + totalOffset,
+          },
+          end: currentStyle.offset + (currentStyle.length - 1),
         });
+      });
 
-        // check for entityRanges starting and add if existing
-        const entitiesStartAtChar = block.entityRanges.filter(range => range.offset === index);
+      // check for entityRanges starting and add if existing
+      const entitiesStartAtChar = block.entityRanges.filter(range => range.offset === index);
 
-        entitiesStartAtChar.forEach(entity => {
-          newText += getEntityStart(raw.entityMap[entity.key]);
-        });
+      entitiesStartAtChar.forEach(entity => {
+        newText += getEntityStart(raw.entityMap[entity.key]);
+      });
 
-        // add the current character to the md string
-        newText += currentChar !== '\n' ? currentChar : '';
+      // add the current character to the md string
+      newText += currentChar !== '\n' ? currentChar : '';
 
-        // check for entityRanges ending and add if existing
-        const entitiesEndAtChar = block.entityRanges.filter(
-          range => range.offset + range.length - 1 === index,
-        );
+      // check for entityRanges ending and add if existing
+      const entitiesEndAtChar = block.entityRanges.filter(
+        range => range.offset + range.length - 1 === index,
+      );
 
-        entitiesEndAtChar.forEach(entity => {
-          newText += getEntityEnd(raw.entityMap[entity.key]);
-        });
+      entitiesEndAtChar.forEach(entity => {
+        newText += getEntityEnd(raw.entityMap[entity.key]);
+      });
 
-        // apply the 'ending' tags for any styles that end in the current position in order (stack)
-        while (
-          appliedStyles.length !== 0 &&
-          appliedStyles[appliedStyles.length - 1].end === index
-        ) {
-          const endingStyle = appliedStyles.pop();
+      // apply the 'ending' tags for any styles that end in the current position in order (stack)
+      while (appliedStyles.length !== 0 && appliedStyles[appliedStyles.length - 1].end === index) {
+        const endingStyle = appliedStyles.pop();
 
-          newText += endingStyle.symbol;
+        newText += endingStyle.symbol;
 
-          newText = fixWhitespacesInsideStyle(newText, endingStyle);
-          totalOffset += endingStyle.symbol.length;
-        }
+        newText = fixWhitespacesInsideStyle(newText, endingStyle);
+        totalOffset += endingStyle.symbol.length;
+      }
 
-        return newText;
-      }, '');
+      return newText;
+    }, '');
 
-      returnString = applyWrappingBlockStyle(block.type, returnString);
-      returnString = applyAtomicStyle(block, raw.entityMap, returnString);
-    });
+    returnString = applyWrappingBlockStyle(block.type, returnString);
+    returnString = applyAtomicStyle(block, raw.entityMap, returnString);
+  });
 
   return returnString;
 }
