+++ src/client/helpers/editorHelper.js
@@ -1,16 +1,6 @@
 import uuidv4 from 'uuid/v4';
-import { last, find, has, uniqWith, isEqual, differenceWith, head, get, keyBy } from 'lodash';
-
-import { CompositeDecorator } from 'draft-js';
-import { Entity, findLinkEntities } from '../components/EditorExtended';
-import ObjectLink, {
-  findObjEntities,
-} from '../components/EditorExtended/components/entities/objectlink';
-import Link from '../components/EditorExtended/components/entities/link';
-import ImageSideButton from '../components/EditorExtended/components/sides/ImageSideButton';
-import VideoSideButton from '../components/EditorExtended/components/sides/VideoSideButton';
-import ObjectSideButton from '../components/EditorExtended/components/sides/ObjectSideButton';
-import SeparatorButton from '../components/EditorExtended/components/sides/SeparatorSideButton';
+import { differenceWith, find, get, has, head, isEqual, keyBy, last, uniqWith } from 'lodash';
+import { Entity } from '../components/EditorExtended';
 
 export const EDITOR_ACTION_ADD = 'add';
 export const EDITOR_ACTION_REMOVE = 'remove';
@@ -55,8 +45,8 @@ export const getLinkedObjects = contentStateRaw => {
 export const getReviewTitle = (campaignData, linkedObjects) => {
   const firstTitle = get(campaignData, 'requiredObject.name', '');
   const secondTitle = get(campaignData, 'secondaryObject.name', '');
-  const requiredObj = get(linkedObjects, '[0]', '');
-  const secondObj = get(linkedObjects, '[1]', '');
+  const requiredObj = get(linkedObjects, '[0]', {});
+  const secondObj = get(linkedObjects, '[1]', {});
   const reviewTitle = `Review: ${firstTitle}, ${secondTitle}`;
 
   const topics = [];
@@ -112,36 +102,6 @@ export const getCurrentLoadObjects = (response, value) => {
 
   return loadObjects;
 };
-
-export const defaultDecorators = new CompositeDecorator([
-  {
-    strategy: findObjEntities,
-    component: ObjectLink,
-  },
-  {
-    strategy: findLinkEntities,
-    component: Link,
-  },
-]);
-
-export const SIDE_BUTTONS = [
-  {
-    title: 'Image',
-    component: ImageSideButton,
-  },
-  {
-    title: 'Video',
-    component: VideoSideButton,
-  },
-  {
-    title: 'Object',
-    component: ObjectSideButton,
-  },
-  {
-    title: 'Separator',
-    component: SeparatorButton,
-  },
-];
 
 const getDifferOfContents = (iteratedRowContent, rowContent) =>
   iteratedRowContent.filter((object, index) => !isEqual(object, rowContent[index]));
